[FEATURE] #38 Unify csv axis assignment for doughnut/pie charts
Adapt the presentation of doughnut and pie charts so that multiple rows
in the csv will be displayed as multiple pies/doughnuts. This is the
format that is already in use for line, bar and other chart types.
A csv with a single data row will now result in a single pie/doughnut.